Moved core.trace to physics.quantum.trace
