+++ sympy/physics/quantum/density.py
@@ -1,13 +1,13 @@
 from itertools import product
 
 from sympy import Tuple, Add, Mul, Matrix, log, expand, S
-from sympy.core.trace import Tr
 from sympy.printing.pretty.stringpict import prettyForm
 from sympy.physics.quantum.dagger import Dagger
 from sympy.physics.quantum.operator import HermitianOperator
 from sympy.physics.quantum.represent import represent
 from sympy.physics.quantum.matrixutils import numpy_ndarray, scipy_sparse_matrix, to_numpy
 from sympy.physics.quantum.tensorproduct import TensorProduct, tensor_product_simp
+from sympy.physics.quantum.trace import Tr
 
 
 class Density(HermitianOperator):
