+++ sympy/physics/quantum/tests/test_density.py
@@ -1,5 +1,4 @@
 from sympy import symbols, S, log, Rational
-from sympy.core.trace import Tr
 from sympy.external import import_module
 from sympy.physics.quantum.density import Density, entropy, fidelity
 from sympy.physics.quantum.state import Ket, TimeDepKet
@@ -9,6 +8,7 @@ from sympy.physics.quantum.dagger import Dagger
 from sympy.physics.quantum.cartesian import XKet, PxKet, PxOp, XOp
 from sympy.physics.quantum.spin import JzKet
 from sympy.physics.quantum.operator import OuterProduct
+from sympy.physics.quantum.trace import Tr
 from sympy.functions import sqrt
 from sympy.testing.pytest import raises
 from sympy.physics.quantum.matrixutils import scipy_sparse_matrix
