+++ sympy/physics/quantum/tests/test_operator.py
@@ -10,8 +10,8 @@ from sympy.physics.quantum.operator import (Operator, UnitaryOperator,
 from sympy.physics.quantum.state import Ket, Bra, Wavefunction
 from sympy.physics.quantum.qapply import qapply
 from sympy.physics.quantum.represent import represent
-from sympy.core.trace import Tr
 from sympy.physics.quantum.spin import JzKet, JzBra
+from sympy.physics.quantum.trace import Tr
 from sympy.matrices import eye
