+++ sympy/physics/quantum/tests/test_tensorproduct.py
@@ -10,7 +10,7 @@ from sympy.physics.quantum.dagger import Dagger
 from sympy.physics.quantum.qubit import Qubit, QubitBra
 from sympy.physics.quantum.operator import OuterProduct
 from sympy.physics.quantum.density import Density
-from sympy.core.trace import Tr
+from sympy.physics.quantum.trace import Tr
 
 A, B, C, D = symbols('A,B,C,D', commutative=False)
 x = symbols('x')
