+++ sympy/physics/quantum/tests/test_trace.py
@@ -1,5 +1,5 @@
 from sympy import symbols, Matrix, Tuple
-from sympy.core.trace import Tr
+from sympy.physics.quantum.trace import Tr
 from sympy.testing.pytest import raises
